+++ ec/src/msm/variable_base/mod.rs
@@ -6,10 +6,33 @@ use crate::{AffineCurve, ProjectiveCurve};
 #[cfg(feature = "parallel")]
 use rayon::prelude::*;
 
-pub struct VariableBaseMSM;
+pub mod stream_pippenger;
+pub use stream_pippenger::*;
 
-impl VariableBaseMSM {
-    pub fn multi_scalar_mul<G: AffineCurve>(
+pub struct VariableBase;
+
+impl VariableBase {
+    /// Optimized implementation of multi-scalar multiplication.
+    ///
+    /// Will return `None` if `bases` and `scalar` have different lengths.
+    ///
+    /// Reference: [`VariableBase::msm`]
+    pub fn msm_checked_len<G: AffineCurve>(
+        bases: &[G],
+        scalars: &[<G::ScalarField as PrimeField>::BigInt],
+    ) -> Option<G::Projective> {
+        (bases.len() == scalars.len()).then(|| Self::msm(bases, scalars))
+    }
+
+    /// Optimized implementation of multi-scalar multiplication.
+    ///
+    /// Will multiply the tuples of the diagonal product of `bases × scalars`
+    /// and sum the resulting set. Will iterate only for the elements of the
+    /// smallest of the two sets, ignoring the remaining elements of the biggest
+    /// set.
+    ///
+    /// ∑i (Bi · Si)
+    pub fn msm<G: AffineCurve>(
         bases: &[G],
         scalars: &[<G::ScalarField as PrimeField>::BigInt],
     ) -> G::Projective {