+++ ff/build.rs
@@ -1,5 +1,5 @@
 extern crate rustc_version;
-use rustc_version::{version, version_meta, Channel, Version};
+use rustc_version::{version_meta, Channel};
 
 fn main() {
     println!("cargo:rerun-if-changed=build.rs");
@@ -17,11 +17,5 @@ fn main() {
     let should_use_asm = asm_enabled && bmi2_enabled && adx_enabled && is_x86_64 && is_nightly;
     if should_use_asm {
         println!("cargo:rustc-cfg=use_asm");
-    }
-
-    // TODO: remove this once RFC 2495 ships
-    if version().expect("Installed rustc version unparseable!") < Version::parse("1.51.0").unwrap()
-    {
-        panic!("This code base uses const generics and requires a Rust compiler version greater or equal to 1.51.0");
     }
 }