+++ ff/src/fields/models/cubic_extension.rs
@@ -6,7 +6,6 @@ use ark_std::{
     cmp::{Ord, Ordering, PartialOrd},
     fmt,
     io::{Read, Result as IoResult, Write},
-    marker::PhantomData,
     ops::{Add, AddAssign, Div, DivAssign, Mul, MulAssign, Neg, Sub, SubAssign},
     vec::Vec,
 };
@@ -25,10 +24,15 @@ use crate::{
     ToConstraintField, UniformRand,
 };
 
+/// Defines a Cubic extension field from a cubic non-residue.
 pub trait CubicExtParameters: 'static + Send + Sync {
     /// The prime field that this cubic extension is eventually an extension of.
     type BasePrimeField: PrimeField;
     /// The base field that this field is a cubic extension of.
+    ///
+    /// Note: while for simple instances of cubic extensions such as `Fp3`
+    /// we might see `BaseField == BasePrimeField`, it won't always hold true.
+    /// E.g. for an extension tower: `BasePrimeField == Fp`, but `BaseField == Fp2`.
     type BaseField: Field<BasePrimeField = Self::BasePrimeField>;
     /// The type of the coefficients for an efficient implemntation of the
     /// Frobenius endomorphism.
@@ -60,6 +64,8 @@ pub trait CubicExtParameters: 'static + Send + Sync {
     );
 }
 
+/// An element of a cubic extension field F_p\[X\]/(X^3 - P::NONRESIDUE) is
+/// represented as c0 + c1 * X + c2 * X^2, for c0, c1, c2 in `P::BaseField`.
 #[derive(Derivative)]
 #[derivative(
     Default(bound = "P: CubicExtParameters"),
@@ -74,19 +80,31 @@ pub struct CubicExtField<P: CubicExtParameters> {
     pub c0: P::BaseField,
     pub c1: P::BaseField,
     pub c2: P::BaseField,
-    #[derivative(Debug = "ignore")]
-    #[doc(hidden)]
-    pub _parameters: PhantomData<P>,
 }
 
 impl<P: CubicExtParameters> CubicExtField<P> {
+    /// Create a new field element from coefficients `c0`, `c1` and `c2`
+    /// so that the result is of the form `c0 + c1 * X + c2 * X^2`.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// # use ark_std::test_rng;
+    /// # use ark_test_curves::bls12_381::{Fq2 as Fp2, Fq6 as Fp6};
+    /// # use ark_test_curves::bls12_381::Fq6Parameters;
+    /// # use ark_std::UniformRand;
+    /// # use ark_ff::models::fp6_3over2::Fp6ParamsWrapper;
+    /// use ark_ff::models::cubic_extension::CubicExtField;
+    ///
+    /// let c0: Fp2 = Fp2::rand(&mut test_rng());
+    /// let c1: Fp2 = Fp2::rand(&mut test_rng());
+    /// let c2: Fp2 = Fp2::rand(&mut test_rng());
+    /// # type Params = Fp6ParamsWrapper<Fq6Parameters>;
+    /// // `Fp6` a degree-3 extension over `Fp2`.
+    /// let c: CubicExtField<Params> = Fp6::new(c0, c1, c2);
+    /// ```
     pub fn new(c0: P::BaseField, c1: P::BaseField, c2: P::BaseField) -> Self {
-        CubicExtField {
-            c0,
-            c1,
-            c2,
-            _parameters: PhantomData,
-        }
+        Self { c0, c1, c2 }
     }
 
     pub fn mul_assign_by_base_field(&mut self, value: &P::BaseField) {
@@ -95,12 +113,19 @@ impl<P: CubicExtParameters> CubicExtField<P> {
         self.c2.mul_assign(value);
     }
 
-    /// Calculate the norm of an element with respect to the base field `P::BaseField`.
+    /// Calculate the norm of an element with respect to the base field
+    /// `P::BaseField`. The norm maps an element `a` in the extension field
+    /// `Fq^m` to an element in the BaseField `Fq`.
+    /// `Norm(a) = a * a^q * a^(q^2)`
     pub fn norm(&self) -> P::BaseField {
+        // w.r.t to BaseField, we need the 0th, 1st & 2nd powers of `q`
+        // Since Frobenius coefficients on the towered extensions are
+        // indexed w.r.t. to BasePrimeField, we need to calculate the correct index.
+        let index_multiplier = P::BaseField::extension_degree() as usize;
         let mut self_to_p = *self;
-        self_to_p.frobenius_map(1);
+        self_to_p.frobenius_map(1 * index_multiplier);
         let mut self_to_p2 = *self;
-        self_to_p2.frobenius_map(2);
+        self_to_p2.frobenius_map(2 * index_multiplier);
         self_to_p *= &(self_to_p2 * self);
         assert!(self_to_p.c1.is_zero() && self_to_p.c2.is_zero());
         self_to_p.c0
@@ -109,12 +134,11 @@ impl<P: CubicExtParameters> CubicExtField<P> {
 
 impl<P: CubicExtParameters> Zero for CubicExtField<P> {
     fn zero() -> Self {
-        CubicExtField {
-            c0: P::BaseField::zero(),
-            c1: P::BaseField::zero(),
-            c2: P::BaseField::zero(),
-            _parameters: PhantomData,
-        }
+        Self::new(
+            P::BaseField::zero(),
+            P::BaseField::zero(),
+            P::BaseField::zero(),
+        )
     }
 
     fn is_zero(&self) -> bool {
@@ -124,12 +148,11 @@ impl<P: CubicExtParameters> Zero for CubicExtField<P> {
 
 impl<P: CubicExtParameters> One for CubicExtField<P> {
     fn one() -> Self {
-        CubicExtField {
-            c0: P::BaseField::one(),
-            c1: P::BaseField::zero(),
-            c2: P::BaseField::zero(),
-            _parameters: PhantomData,
-        }
+        Self::new(
+            P::BaseField::one(),
+            P::BaseField::zero(),
+            P::BaseField::zero(),
+        )
     }
 
     fn is_one(&self) -> bool {
@@ -630,8 +653,24 @@ where
 #[cfg(test)]
 mod cube_ext_tests {
     use super::*;
-    use crate::test_field::{Fq, Fq2, Fq6};
     use ark_std::test_rng;
+    use ark_test_curves::{
+        bls12_381::{Fq, Fq2, Fq6},
+        mnt6_753::Fq3,
+        Field,
+    };
+
+    #[test]
+    fn test_norm_for_towers() {
+        // First, test the simple fp3
+        let mut rng = test_rng();
+        let a: Fq3 = rng.gen();
+        let _ = a.norm();
+
+        // then also the tower 3_over_2, norm should work
+        let a: Fq6 = rng.gen();
+        let _ = a.norm();
+    }
 
     #[test]
     fn test_from_base_prime_field_elements() {