+++ poly/src/lib.rs
@@ -1,7 +1,13 @@
 //! This crate implements functions for manipulating polynomials over finite
 //! fields, including FFTs.
 #![cfg_attr(not(feature = "std"), no_std)]
-#![warn(unused, future_incompatible, nonstandard_style, rust_2018_idioms)]
+#![warn(
+    unused,
+    future_incompatible,
+    nonstandard_style,
+    rust_2018_idioms,
+    rust_2021_compatibility
+)]
 #![forbid(unsafe_code)]
 #![allow(
     clippy::many_single_char_names,
@@ -23,10 +29,12 @@ pub mod polynomial;
 pub use domain::{
     EvaluationDomain, GeneralEvaluationDomain, MixedRadixEvaluationDomain, Radix2EvaluationDomain,
 };
-pub use evaluations::multivariate::multilinear::{
-    DenseMultilinearExtension, MultilinearExtension, SparseMultilinearExtension,
+pub use evaluations::{
+    multivariate::multilinear::{
+        DenseMultilinearExtension, MultilinearExtension, SparseMultilinearExtension,
+    },
+    univariate::Evaluations,
 };
-pub use evaluations::univariate::Evaluations;
 pub use polynomial::{multivariate, univariate, MVPolynomial, Polynomial, UVPolynomial};
 
 #[cfg(test)]