+++ test-curves/src/mnt6_753/fr.rs
@@ -1,0 +1,1 @@
+pub use crate::mnt4_753::{Fq as Fr, FqParameters as FrParameters};