+++ test-templates/src/fields.rs
@@ -2,8 +2,7 @@
 #![allow(clippy::eq_op)]
 use ark_ff::fields::{FftField, FftParameters, Field, LegendreSymbol, PrimeField, SquareRootField};
 use ark_serialize::{buffer_bit_byte_size, Flags, SWFlags};
-use ark_std::io::Cursor;
-use ark_std::rand::Rng;
+use ark_std::{io::Cursor, rand::Rng};
 
 pub const ITERATIONS: u32 = 40;